massive/pyproject.toml


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "massive-automator"
version = "0.1.0"
description = "Batch-generate Massive-style presets, validate JSON patches, and push macros via MIDI."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.0",
  "mido>=1.3.2",
  "python-rtmidi>=1.5.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "flake8>=7.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12.20240808",
]

[project.scripts]
massive = "massive.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["massive*"]

[tool.flake8]
max-line-length = 100
extend-exclude = ["__pycache__", ".venv", "venv", "build", "dist"]
per-file-ignores = []

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
strict_equality = true

[tool.pytest.ini_options]
addopts = "-q -ra"
testpaths = ["tests"]